Create PauseMenuController
The variable gameIsPaused is actually a scriptable object so I can pass the information of the game being paused between my scripts. I had to do this because the slow-motion effect would unpause the game whenever you held down shift, but we wanted time to still stay paused. Check my GameState script for that scriptable object and other ones.